+++ spring-integration-jpa/src/main/java/org/springframework/integration/jpa/support/parametersource/ExpressionEvaluatingParameterSourceFactory.java
@@ -16,10 +16,17 @@
 package org.springframework.integration.jpa.support.parametersource;
 
 import java.util.ArrayList;
-import java.util.Collections;
+import java.util.Collection;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 
 import org.springframework.beans.factory.BeanFactory;
+import org.springframework.expression.Expression;
+import org.springframework.expression.ExpressionException;
 import org.springframework.integration.jpa.support.JpaParameter;
 import org.springframework.integration.jpa.support.parametersource.ExpressionEvaluatingParameterSourceUtils.ParameterExpressionEvaluator;
 import org.springframework.util.Assert;
@@ -28,10 +35,15 @@ import org.springframework.util.Assert;
  *
  * @author Gunnar Hillert
  * @author Gary Russell
+ * @author Artem Bilan
  * @since 2.2
  *
  */
 public class ExpressionEvaluatingParameterSourceFactory implements ParameterSourceFactory {
+
+	private final static Log logger = LogFactory.getLog(ExpressionEvaluatingParameterSourceFactory.class);
+
+	private static final Object ERROR = new Object();
 
 	private volatile List<JpaParameter> parameters;
 
@@ -42,7 +54,7 @@ public class ExpressionEvaluatingParameterSourceFactory implements ParameterSour
 	}
 
 	public ExpressionEvaluatingParameterSourceFactory(BeanFactory beanFactory) {
-		this.parameters = Collections.unmodifiableList(new ArrayList<JpaParameter>());
+		this.parameters = new ArrayList<JpaParameter>();
 		this.expressionEvaluator.setBeanFactory(beanFactory);
 	}
 
@@ -67,6 +79,120 @@ public class ExpressionEvaluatingParameterSourceFactory implements ParameterSour
 
 	public PositionSupportingParameterSource createParameterSource(final Object input) {
 		return new ExpressionEvaluatingParameterSource(input, this.parameters, expressionEvaluator);
+	}
+
+
+	class ExpressionEvaluatingParameterSource implements PositionSupportingParameterSource {
+
+		private final Object input;
+
+		private volatile Map<String, Object> values = new HashMap<String, Object>();
+
+		private final List<JpaParameter> parameters;
+
+		private final Map<String, JpaParameter> parametersMap;
+
+		private final ParameterExpressionEvaluator expressionEvaluator;
+
+		ExpressionEvaluatingParameterSource(Object input, List<JpaParameter> parameters, ParameterExpressionEvaluator expressionEvaluator) {
+
+			this.input = input;
+			this.expressionEvaluator = expressionEvaluator;
+			this.parameters = parameters;
+			this.parametersMap = new HashMap<String, JpaParameter>(parameters.size());
+			for (JpaParameter parameter : parameters) {
+				this.parametersMap.put(parameter.getName(), parameter);
+			}
+			this.values.putAll(ExpressionEvaluatingParameterSourceUtils.convertStaticParameters(parameters));
+
+		}
+
+		public Object getValueByPosition(int position) {
+
+			Assert.isTrue(position >= 0, "The position must be be non-negative.");
+
+			if (position <= parameters.size()) {
+
+				final JpaParameter parameter = parameters.get(position);
+
+				if (parameter.getValue() != null) {
+					return parameter.getValue();
+				}
+
+				if (parameter.getExpression() != null) {
+					Expression expression;
+
+					if (input instanceof Collection<?>) {
+						expression = parameter.getProjectionExpression();
+					}
+					else {
+						expression = parameter.getSpelExpression();
+					}
+
+					final Object value = this.expressionEvaluator.evaluateExpression(expression, input);
+					if (parameter.getName() != null) {
+						values.put(parameter.getName(), value);
+					}
+					if (logger.isDebugEnabled()) {
+						logger.debug("Resolved expression " + expression + " to " + value);
+					}
+					return value;
+
+				}
+
+			}
+
+			return null;
+
+		}
+
+		public Object getValue(String paramName) {
+			if (values.containsKey(paramName)) {
+				return values.get(paramName);
+			}
+
+			if (!this.parametersMap.containsKey(paramName)) {
+				JpaParameter parameter = new JpaParameter(paramName, null, paramName);
+				ExpressionEvaluatingParameterSourceFactory.this.parameters.add(parameter);
+				this.parametersMap.put(paramName, parameter);
+			}
+
+			JpaParameter jpaParameter = this.parametersMap.get(paramName);
+
+			Expression expression = null;
+
+			if (input instanceof Collection<?>) {
+				expression = jpaParameter.getProjectionExpression();
+			}
+			else {
+				expression = jpaParameter.getSpelExpression();
+			}
+
+			final Object value = this.expressionEvaluator.evaluateExpression(expression, input);
+			values.put(paramName, value);
+			if (logger.isDebugEnabled()) {
+				logger.debug("Resolved expression " + expression + " to " + value);
+			}
+			return value;
+		}
+
+		public boolean hasValue(String paramName) {
+			try {
+				final Object value = getValue(paramName);
+				if (value == ERROR) {
+					return false;
+				}
+			}
+			catch (ExpressionException e) {
+				if (logger.isDebugEnabled()) {
+					logger.debug("Could not evaluate expression", e);
+				}
+				values.put(paramName, ERROR);
+				return false;
+			}
+			return true;
+		}
+
 	}
 
 }
