+++ spring-integration-jpa/src/main/java/org/springframework/integration/jpa/support/parametersource/ExpressionEvaluatingParameterSourceUtils.java
@@ -65,6 +65,10 @@ class ExpressionEvaluatingParameterSourceUtils {
 		return staticParameters;
 	}
 
+	/**
+	 * Simple {@link AbstractExpressionEvaluator} implementation
+	 * to increase the visibility of protected methods.
+	 */
 	public static class ParameterExpressionEvaluator extends AbstractExpressionEvaluator {
 
 		@Override
